+++ Runtime/AbstractNodes/RandomNode.cs
@@ -31,20 +31,20 @@ namespace Voxell.Rasa
     {
       base.OnEnable();
       // connection of other ports to the data port at the random node
-      if (!fieldNames.Contains("list"))
+      if (!fieldNames.Contains("array"))
       {
-        fieldNames.Add("list");
+        fieldNames.Add("array");
         connections.Add(new Connection());
       }
     }
 
-    protected override void OnStart()
+    protected override void OnStart(ref RasaNLP rasaNLP)
     {
       rasaState = RasaState.Running;
       Connection connection = connections[0];
-      List<T> list = (List<T>)connection.GetValue(0);
-      int selectionIdx = Random.Range(0, list.Count);
-      selection = list[selectionIdx];
+      T[] array = (T[])connection.GetValue(0);
+      int selectionIdx = Random.Range(0, array.Length);
+      selection = array[selectionIdx];
     }
     protected override RasaState OnUpdate() => RasaState.Success;
     protected override void OnStop() {}