+++ Runtime/Nodes/MaximumEntropyTokenizerNode.cs
@@ -19,24 +19,25 @@ All rights reserved.
 
 using System;
 using System.Collections.Generic;
+using Voxell.NLP.Tokenize;
 
 namespace Voxell.Rasa
 {
-  public class RandomStringNode : RandomNode<string>
+  public class MaximumEntropyTokenizerNode : ArrayNode<string>
   {
-    new public static string pathName = "Random/Random String";
+    new public static string pathName = "NLP/Maximum Entropy Tokenizer";
 
     public override List<PortInfo> CreateInputPorts()
     {
       List<PortInfo> portInfos = base.CreateInputPorts();
-      portInfos.Add(new PortInfo(CapacityInfo.Single, typeof(List<string>), "array", EdgeColor.strList));
+      portInfos.Add(new PortInfo(CapacityInfo.Single, typeof(string), "data", EdgeColor.str));
       return portInfos;
     }
 
     public override List<PortInfo> CreateOutputPorts()
     {
       List<PortInfo> portInfos = base.CreateOutputPorts();
-      portInfos.Add(new PortInfo(CapacityInfo.Multi, typeof(string), "selection", EdgeColor.str));
+      portInfos.Add(new PortInfo(CapacityInfo.Multi, typeof(List<string>), "array", EdgeColor.strList));
       return portInfos;
     }
 
@@ -57,6 +58,14 @@ namespace Voxell.Rasa
         connections[0].RemoveAt(0);
         return true;
       }
+    }
+
+    protected override void OnStart(ref RasaNLP rasaNLP)
+    {
+      rasaState = RasaState.Running;
+      Connection connection = connections[0];
+      string text = (string)connection.GetValue(0);
+      array = rasaNLP.maximumEntropyTokenizer.Tokenize(text);
     }
   }
 }