+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [1.1.0]
+
+- New Features
+  - Tokenizer Node: tokenize sentence *(string)* into array of tokens *(strings)*
+
 ## [1.0.0]
 
 - Initial release.
+- Foundation of node graph system *(edit, save, create, etc.)* and basic abstract nodes.